import saas request overrides module into worker runtime
move worker main into its own module to avoid dependency conflicts
